[FE]Rename: paths config 수정 및 components내 login 폴더 Login으로 수정
