+++ frontend/src/pages/Auth/LoginPage.tsx
@@ -1,8 +1,8 @@
 import styled from 'styled-components';
 
-import Entry from '@/components/login/Entry';
+import Entry from '@/components/Login/Entry';
 import { useEffect, useState } from 'react';
-import Form from '@/components/login/Form';
+import Form from '@/components/Login/Form';
 
 function LoginPage() {
   //TODO submit 함수 자식한테 뿌려주기 통신 로직만 page 컴포넌트에서
